order feature start

+++ src/main/java/com/orange/model/Item.java
@@ -1,7 +1,5 @@
 package com.orange.model;
 
-import java.math.BigDecimal;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -23,7 +21,7 @@ public class Item {
 	private Long itemId;
 	@NonNull private String category;
 	@NonNull private String name;
-	@NonNull private BigDecimal price;
+	@NonNull private int price;
 	@NonNull private int quantity;
 	@NonNull private int sales;
 	@NonNull private String description;
